Removed interface SiteManager
